backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.11.0
python-multipart>=0.0.12
openpyxl>=3.1.2
pandas>=2.2.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
